Refactor type checking of user-set variables
Previously, the functions which checked that user-set configuration
variables were of the correct type both checked the type of the
variable, and printed an error message if the type was incorrect. In
retrospect, this is a mixing of responsibilities. Now, the functions
just check the type of the variable passed and return a boolean. At the
call site the value of the boolean is used to determine whether an error
message is printed.

Additionally, the functions were previously autoloaded. However, because
they was called in `plugin/topiary.vim`, `autoload/topiary.vim` was
sourced when Vim started, defeating the purpose of autoloading. Now the
functions have been moved to `plugin/topiary.vim` instead of being
autoloaded.